requirements.txt
typer
SQLAlchemy
rich
docker
requests
PyYAML
python-dotenv
tenacity
openai
google-cloud-aiplatform
gitpython
prometheus-client
psutil
platformdirs
